fix: cast vectorizer output to float32 for MLP

+++ app.py
@@ -51,7 +51,7 @@ def _train():
                     X_texts.append(normalize(ex))
                     y_labels.append(intent)
         vec = CountVectorizer(analyzer="char_wb", ngram_range=(3, 5))
-        X_v = vec.fit_transform(X_texts)
+        X_v = vec.fit_transform(X_texts).astype("float32")
         clf = MLPClassifier(hidden_layer_sizes=(64,), max_iter=200, random_state=42, early_stopping=True)
         clf.fit(X_v, y_labels)
         _vec, _X_v, _clf, _cfg = vec, X_v, clf, cfg
